First Pass at Modularizing
Why these changes are being introduced:
The first pass through made it pretty obvious that this should be
handled with an internal module. This is the first pass at setting up
the module version of this code. The existing ECR repos were not
touched.

How this addresses that need:
* Create embedded ecr module (in /modules/ecr)
* Create two module calls for fake ECR Repositories to verify that
there are no conflicts when the module is used more than once

Side effects of this change:
None. This is just a test pass before working on moving the existing
ECR repos into the modularized version.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/ENSY-77
* https://mitlibraries.atlassian.net/browse/ENSY-78