Clean up for PR and merge
Why these changes are being introduced:
Some necessary cleanup of documentation and tagging of resources.

How this addresses that need:
* Create tags for all taggable resources
* Clean up root README
* Clean up modules README
* Clean up in-line documentation in the module
* Clean up in-line documentation in the root module
* Remove sample repo modules
* Prep for shift to modules for alma-hook-lambdas and ppod

Side effects of this change:
None. This version will not be deployed -- this is just cleanup before
running a formal `terraform destroy` to remove the original ECR repos in
preparation for redeploying the repos via the embedded module.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/ENSY-77
* https://mitlibraries.atlassian.net/browse/ENSY-78