add timdex related ECR's
# Subject
Adds the timdex related ECR's

# Why these changes are being introduced:
This adds the timdex related ECR's so that the timdex stepfunction and related lambda and ECS tasks can collapse into one repo.

# How this addresses that need:
Creates the ECR's and, using the module, the OIDC role used in the github repo to update the ECR image.

# Side effects of this change:
The timdex-stepfunction and timdex-infrastructure repos need to be refactored.