Add Makefile and caller workflow outputs
Why these changes are being introduced:
This started as a simple task to move the Makefile and caller workflow
text from the Fargate/Lambda repos back one step to the repo that
generates the ECR repositories. It turned into a small refactor in that
process.

This repo now generates the app repo Makefile (for deploys to Dev1)
and the caller workflows for dav & stage builds and prod promotes.

How this addresses that need:
* Move each module call into its own .tf file
* Add Tf outputs to each .tf file that generates a new ECR repository
* Add the necessary template files to the `files/` directory to format
the text for the Makefiles and caller workflows
* Add outputs to the embedded ece module to provide the text to
insert into the template files when generating the Tf outputs

Side effects of this change:
None. When this is deployed, it will just generate a long list of
new Tf outputs.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-518