Complete Outputs for Automation
Why these changes are being introduced:
This is the final round of testing and cleanup for enabling the correct
Terraform outputs for dependent Fargate/Lambda application repositories
and automating build/push/deploy/promote process.

After conversations with EngX, there was a desire for an "easy" way
for devs to be able to build/deploy a container/Lambda to stage (and
eventually prod) in an emergency. This handles the first phase of that:
providing instructions and make commands that a dev can use, if needed,
to build and push directly to Stage-Workloads.

How this addresses that need:
* Clean up resource naming for alma-webhook-lambdas application
* Clean up outputs for the ppod application prod_promote workflow
* Create all the outputs for all the ECRs related to the timdex pipeline
applications (mario, oaiharvester, & transmogrifier)
* Fix naming for the existing outputs for the timdex-pipeline-lambdas
application
* Add repo_name as an additional output for the module (so we can
easily reference the repo_name input that was passed to the module
when we need it for the associated Terraform outputs)
* Add the app_repo name to the tags for the ssm_outputs generated by
the embedded module
* Clean up Fargate-related templates comments and naming
* Clean up Lambda-related templates comments and naming
* Fix variable name mistake in the fargate-prod-promote workflow
template
* Update all caller workflow templates to reference the `main` branch
in the shared .github repository for the shared workflows
* Add instructions to the Makefile on how to push directly to stage
* Add additional make commands that can build/push a container/Lambda
to Stage-Workloads (but these depend on the dev manually setting a
few local CLI environment variables)
* Update README

Side effects of this change:
This will not make any changes to the existing ECR repositories -- it
will just update the Terraform outputs.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-518