Add GIS-related ECRs
Why these changes are being introduced:
In preparation for the migration of the GIS stack to AWS Org, this
creates the ECR repositories for the various containers. This also
updates Terraform and AWS Provider version constraints.

How this addresses that need:
* Creates ECR repositories for GeoServer, GeoWeb, GeoSolr, and
Slingshot
* Updates Terraform version constraint to ~> 1.2
* Updates AWS Provider version constraint to ~> 4.0
* Update typos in variables.tf and in the README

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-501

Resolves: #17 (GH issue)