Rename GIS-related ECRs
Per discussion in the #engineering Slack channel, we
are reverting to using the `docker-` prefix for
our repos that are really just modifications of official 3rd party
Docker containers

Side effects:
This will destory and recreate the ECRs and associated
policies and outputs. This will
require a redeploy of the mitlib-tf-workloads-gis repo
as well as re-uploading the containers for each ECR repository.