Rename Matomo ECR
Why these changes are being introduced:
Per a conversation in the #engineering channel in Slack, we have
decided that all of our repos that are really just modifications of
a third party Docker container should have "docker-" as the prefix
of the name. Since there is a tight relationship between the app
GitHub repo name, the ECR repo name, and the OIDC roles/policies for
automated deployments, the ECR configuration needs to be updated.

How this addresses that need:
* Change the app-repo name from "matomo" to "docker-matomo" in the
module call that creates the ECR and related bits

Side effects of this change:
This will delete the current ECR, roles, policies, and outputs and
recreate them. The GitHub repo for Matomo will also need to be renamed
and the new Makefile and GHA workflow files will need to be updated.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-487