Update Outputs for Shared Workflows
Why these changes are being introduced:
The Carbon app introduces a few changes to our automated build process.
There are changes in the shared workflows in the MIT Libraries .giuthub
repository that reflect the automated build changes necessary for
carbon. The changes in this repo update the Terraform Outputs to match
the new input requirements of the shared workflows.

As part of addressing the needs of the Carbon app, it made sense to
consolidate the shared workflows by using a few conditional steps. The
changes here continue that idea of consolidation with conditional text.

How this addresses that need:
* Create three new template files for dev-build, stage-build, and
prod-promote caller workflow outputs
* Create one new template file for the Makefile output
* Update the output resources for just the carbon ECR to use the
new template files

Side effects of this change:
None. In fact, the only changes are changes to output values. This will
not make any changes to the infrastructure in AWS.

In the near future, we will need to update the Terraform Outputs to
use the new templates and eventually remove the outdated templates
once we have upgraded across our multiple repositories.