Update wcd2reshare-related Outputs
Why these changes are being introduced:
There are new shared workflows in the shared .github repository for
handling the Fargate/Lambda deployment pipeline.

How this addresses that need:
* Update the wcd2reshare-related output in dev to use the updated
template for the dev-build caller workflow
* Update the wcd2reshare-related output in dev to use the update
template for the Makefile
* Update the wcd2reshare-related output in stage to use the updated
template for the stage-build caller workflow
* Update the wcd2reshare-related output in prod to use the updated
template for the prod-promote caller workflow

Side effects of this change:
None.