Add ECR for Geo-Harvester
Why these changes are being introduced:
A new geo-harvester function is being written for the GDT project to
add geographic and geospatial data to TIMDEX. This creates the ECR repo
in AWS to hold the built containers.

How this addresses that need:
* Add a new ECR to the set of TIMDEX ECRs called geo-harvester
* Update README

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-38