Fix Typo for Browsertrix
Why these changes are being introduced:
The code for the browsertrix app had some misspellings in it (where the
"ws" was sometimes transposed as "sw"). This fixes that without making
any changes to the deployed infrastructure.

How this addresses that need:
* Fix all the misspellings of "browsertrix" in the code
* Create a number of `moved {}` blocks to ensure that after the
misspellings are fixed, no infrastructure related to browsertrix is
changed
* Update README

Side effects of this change:
None.