Remove Geoweb Testing Resources
Why these changes are being introduced:
We have moved beyond trying to migrate the existing geoweb stack from
legacy AWS to our AWS Org. We cleaned up all the other deployed
resources, but we forgot to remove the ECR repositories and associated
resources. This takes care of that.

How this addresses that need:
* Remove all the ECR and ECR-related resources for the failed attempt
to migrate geoweb from legacy AWS to AWS Org

Side effects of this change:
None.