Create alma-sapinvoices-ui ECR Repository
Why these changes are being introduced:
As part of adding a web-based UI to run the SAP Invoices task, we need
to set up an ECR Repository to store the container images that will
be used in the Lambda function.

How this addresses that need:
* Create a standard ECR Repository with permissions and sample text
for GHA workflows and Makefiles for a Lambda function using a
container for its image

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1011

fix for function name