rename asati repository
Why these changes are being introduced:

This commit renames the asati app repository from docker-asati to asati. The
docker prefix is only used for third party applications that release docker
images that the Libraries makes use of.

How this addresses that need:
* application repository renamed
* terraform-docs was updated to the latest version in the pre-commit hook

Side effects of this change:

None

Changes to be committed:
      modified:   .pre-commit-config.yaml
      modified:   asati_ecr.tf