Create ECR for s3-bagit-validator
Why these changes are being introduced:
As part of CDPS and directly related to the Isilon Replacement project,
this starts the build out of a new app for validating AIPs stored in
S3 as part of the Archivematica implementation.

How this addresses that need:
* Create a new set of ECR-related resources for the s3-bagit-validator
application that will be part of CDPS
* Remove the `overwrite = true` lines from the Terraform managed
ssm_outputs (that line is deprecated in the current version of the AWS
Provider)

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-177