Create ECR Repository for CDPS-CURT
Why these changes are being introduced:
There is anew application, `cdps-curt`, that will generate cost and
usage reports for CDPS resources in AWS. It will run as a Fargate task,
so it needs and ECR Repository.

How this addresses that need:
* Create a new ECR repository and associated roles for the cdps-curt
application
* Update the README

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1207