ECR Repository for Archival Packaging Tool (APT)
Why these changes are being introduced:
A new archival packaging tool application is getting built and it will
run as a Fargate task, so it needs an ECR repository.

How this addresses that need:
* Create a new ECR repository along with the associate GHA roles for
automation and the Terraform outputs for the GHA workflows for the
application repository
* Update the pre-commit dependencies
* Update the README

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1221

wip