Update Embedded Module for Multi-Region
Why these changes are being introduced:
We need to ensure that the embedded module can correctly create
resources in multiple regions. Since IAM Roles/Policies are "global" we
only want to create those resources in the default `us-east-1` region.
The ECR Repository is regional, so we create it each time the module is
called based on the submitted provider.

How this addresses that need:
* Add the `aws_region.current` data object so that we can track the
region provided when the module is called
* Add conditionals to all the "global" resources so that they only get
created when the module is called for the us-east-1 region
* Reorganize the resource creation in the embedded module slightly

Side effects of this change:
None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-238