Update Tf Outputs
Why these changes are being introduced:
Now that we have ECR Repositories in multiple regions, we need to
genarate a few additional lines for each workflow and Makefile when we
enable an additional region.

How this addresses that need:
* Create an additional template file for the Makefile add-ins
* Create an additional template file for the dev-build add-ins
* Create an additional template file for the stage-build add-ins
* Create an additional template file for the prod-promote add-ins
* Create a new Tf output for the Makefile add-ins for the bagit Lambda
that is deployed in us-west-2
* Create a new Tf output for the dev-build add-ins for the bagit Lambda
that is deployed in us-west-2
* Create a new Tf output for the stage-build add-ins for the bagit
Lambda that is deployed in us-west-2
* Create a new Tf output for the prod-promote add-ins for the bagit
Lambda that is deployed in us-west-2

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-238