New GHA Templates
Why these changes are being introduced:
The new option to pick a CPU architecture for a container requires new
caller workflow templates and new Makefile commands for building and
deploying a container.

How this addresses that need:
* Create new template files for dev, stage, and prod GHA caller
workflows
* Create new Makefile commands (and remove the almost-never-used make
commands to build and push a container directly to Stage-Workloads)
* Update the outputs for the testing repository

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1448